apps/ml/pyproject.toml
[project]
name = "nitbox-ml"
version = "0.1.0"
description = "NITBox ML microservice"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.7.0",
    "httpx==0.27.0",
    "python-dotenv==1.0.0",
    "pandas==2.2.2",
    "scipy==1.13.0",
    "sqlalchemy==2.0.30",
    "psycopg2-binary==2.9.9",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"
